pollNodeInfoCRDAndUpdatePlugin
